rwkv模型支持插件化
