+++ run_rwkv.bat
@@ -1,7 +1,9 @@
 @echo off
 call settings.bat
+set llm_type=rwkv
+rem  LLM模型类型:rwkv
 :a
-%PYTHON% rwkvAPI.py
+%PYTHON% wenda.py
 goto a
 pause
 exit /b
